add git create, del, and rename
These three were all born of my use of the `git_remote_branch` gem. They
are features of that utility gem and it was about time they were pulled
out into plain shell scripts instead of moving around all over the place
and needing that gem to be installed. This simplifies things greatly.

Luckily that gem has a wonderful "explain" feature which allowed me to
lift the series of commands for each of these particular needs.