Add vim-test and vim-tmux-runner
vim-test is a nice piece of work given the fact it works very well with several
different languages and test frameworks, not just ruby or rspec.
Thankfully it supports Vim Tmux Runner as a method of dispatch, so let's
use that too ...

Vim Tmux Runner is @christoomey's doing. It's ridiculous. It's great.
This is now an opportunity to re-train muscle memory to use it instead of
constantly having a tmux pane open that receives my custom commands.

Thank you @christoomey and @janko-m :)