Add some trello magic
I am an unabashed trello fan. This is no secret.

Often I'll be working on a project and want to head to the project's
board to look at what's out there. Navigating to that board is time that
can be shaved away. If I'm in the project already why not just type `$
trello` and have the browser open at that project?

So let's make it happen, plus one extra little feature - adding a new
card to what is normally a "Backlog" list/swim-lane.

1. This util script wraps the [wrk project] - a command line trello
   client. `brew install wrk-trello` to install and get set up.
2. For this to work it assumes you have a .trello file in your $PWD
   containing the id of the board. Pretty easy to get - it's in your
   board's url.
3. Running `trello` alone will open the board in your browser.
4. Running `trello "Some title"` will create a new card in the left-most
   list.

For now, that's it! I'm sure I'll build upon this.

[wrk project]: https://github.com/blangel/wrk