If `.heroku` file is in CWD, set current heroku account
Because I have several accounts on heroku - mostly for client
applications - I often end up running heroku toolbelt commands with the
wrong account authenticated. This commit contains some updates that look
for a .heroku file and, if it exists, fun the heroku accounts plugin to
set the account to the contents of that file.

Also - include a fh() function that allows me to search my history with fzf.