Updates to work with M1 homebrew
The base dir for homebrew changes for m1 arch macs so, we need to take that into account.

Also - move some of these git-related aliases to functions so that they are not run whenever a shell starts up -- only when the function is invoked.

Also - check for the existence of rbenv before trying to invoke it.